ADD: 翻转链表 / 链表中倒数K个节点 / 合并两个排序链表 / 树的字结构 / 二叉树的镜像 / 顺时针打印矩阵 / 包含 min 函数的栈

+++ test/剑指Offer/013-在O(1)时间删除链表节点_test.go
@@ -16,7 +16,7 @@ func TestDeleteNode(t *testing.T) {
 
 	// 删除尾节点
 	lhs = rhs
-	toBeDeleted = list.NodeAt(uint(len(rhs) - 1))
+	toBeDeleted = list.NodeAt(len(rhs) - 1)
 	list.DeleteNode(toBeDeleted)
 	rhs = list.IntArray()
 	assert.Equal(t, len(lhs)-len(rhs), 1)
